requirements.txt
numpy
tensorflow
pillow
tflearn
opencv-python
Cython==0.26.1
kivy
imagehash
sklearn
pyyaml
